修复关注有礼
