+++ README.md
@@ -1,22 +1,44 @@
-# JD-Script
+
+# JD 关注有礼
+
 ##  目录结构
-    JD-Script/                  #JD-Script仓库
-    |-- BlueCoin                # 东东超市兑换
-    |-- getFollowGifts          # JD关注有礼
-    |-- OpenCard                # JD入会领豆
-    |-- getJDCookie             # JD获取Cookie
-    |   
+    JD-Script/                  #主仓库
+    |-- getFollowGifts                # 主目录
+    |   |-- jd_getFollowGift.py       # 主代码 （必要）
+    |   |-- JDCookies.txt             # 存放JD cookie，一行一个ck
+    |   |-- Readme.md                 # 说明书
+    |   `-- start.sh                  # shell脚本（非必要）
     `-- README.md
-    
-    
+
+
+### `【兼容环境】`
+    1.Python3.6+ 环境
+    2.兼容ios设备软件：Pythonista 3、Pyto(已测试正常跑，其他软件自行测试)   
+    3.Windows exe 
+
+    安装依赖模块 :
+    pip3 install requests
+    执行：
+    python3 jd_getFollowGift.py
+
+
+## `【更新记录】`
+    2021.6.6：（v1.0.0 Beta）
+        * Test
+
+
 ###### [GitHub仓库 https://github.com/curtinlv/JD-Script](https://github.com/curtinlv/JD-Script) 
-###### [Gitee仓库 https://gitee.com/curtinlv/JD-Script](https://gitee.com/curtinlv/JD-Script)
 ###### [TG频道 https://t.me/TopStyle2021](https://t.me/TopStyle2021)
 ###### [TG群 https://t.me/topStyle996](https://t.me/topStyle996)
-###### 公众号【TopStyle】
+###### 关注公众号【TopStyle】
 ![TopStyle](https://gitee.com/curtinlv/img/raw/master/gzhcode.jpg)
-
 # 
-    @Last Time: 2021-05-15
+    @Last Version: v1.0.0 Beta
+
+    @Last Time: 2021-06-06 07:57
 
     @Author: Curtin
+#### **仅以学习交流为主，请勿商业用途、禁止违反国家法律 ，转载请留个名字，谢谢!** 
+
+# End.
+[回到顶部](#readme)
